Moving args to common, making args be able to be in all scripts, setting data back to use where the real parquet files live for CI
